requirements.txt
ansible==11.3.0
ansible-core==2.18.3
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-RESTful==0.2.8
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
passlib==1.7.4
psycopg==3.2.5
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
resolvelib==1.0.1
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
